🚀feat(models) add Balance and Transaction models 🛠️refactor(models) add new fields to User model 🛠️refactor(models) set timestamps to false in Currency model ⚙️chore(migrations) create balances and transactions migrations ⚙️chore(seeders) add User seeder 🛠️refactor(routes) add TestController 🚀feat(database) rename migration files
🚀feat(models) add Balance and Transaction models

Two new models have been added: Balance and Transactions. These models will allow users to view their balances and transactions history.

🛠️refactor(models) add new fields to User model

Additional fields have been added to the User model - uuid and created_at. This change will allow for unique identification of users and better tracking of when a user was created.

🛠️refactor(models) set timestamps to false in Currency model

Timestamps have been set to false in the Currency model to remove the overhead of updating the created_at and updated_at columns every time an instance is created or updated.

⚙️chore(migrations) create balances and transactions migrations

Migrations for the new Balance and Transactions models have been created, which include columns such as amount, type, and status.

⚙️chore(seeders) add User seeder

A new user seeder has been added to populate the users table with sample data. The sample data is loaded from a JSON file.

🛠️refactor(routes) add TestController

A new TestController has been added to the routes, which can be used to test various functionality for development and troubleshooting.

🚀feat(database) rename migration files

Migration files for the currency table and user table have been renamed to match the standard format of Laravel. This change will help to maintain consistency across all migration files.