Refactor codebase to implement User API endpoints
This commit refactors the codebase to implement User API endpoints to fetch a single user with transactions and all users with their transactions. It also includes tests for the implemented endpoints. The `UserService` class abstracts the index and show logic from the controller, which has been refactored to use the `UserService`. The `User` model now has new query scopes to filter users by amount range and search for users by keyword. The `Handler` class now returns a JSON response for ModelNotFoundExceptions. Lastly, the `README` file has been updated with instructions on how to run the project with `docker` using `Laravel Sail`.